authoptions exported fix
