+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,50 +1,6 @@
-import NextAuth, { NextAuthOptions } from 'next-auth'
-import CredentialsProvider from 'next-auth/providers/credentials'
-import { connectDB } from '@/utils/db'
-import { findByEmailAndCode } from '@/controllers/Auth'
+import NextAuth from 'next-auth'
+import authOptions from './authOptions'
 
-const authOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: {
-          label: 'Email',
-          type: 'email',
-        },
-        code: { label: 'Code', type: 'text' },
-      },
-      async authorize(credentials, req) {
-        await connectDB()
+const handler = NextAuth(authOptions)
 
-        const authFound = await findByEmailAndCode(
-          credentials!.email,
-          credentials!.code,
-        )
-        if (!authFound) return null
-        return authFound
-      },
-    }),
-  ],
-  session: {
-    strategy: 'jwt',
-  },
-  pages: {
-    signIn: '/auth/code',
-  },
-
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) token.user = user
-      return token
-    },
-    async session({ session, token }) {
-      session.user = token.user as any
-      return session
-    },
-  },
-}
-
-export const handler = NextAuth(authOptions)
-
-export { handler as GET, handler as POST, authOptions }
+export { handler as GET, handler as POST }
