+++ src/app/api/profile/route.ts
@@ -6,7 +6,7 @@ import {
   updateProfileController,
 } from '@/controllers/Profile'
 import { getServerSession } from 'next-auth'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
+import authOptions from '@/app/api/auth/[...nextauth]/authOptions'
 
 export const GET = async () => {
   const session = await getServerSession(authOptions)
