+++ src/utils/fetch-api.ts
@@ -1,8 +1,8 @@
 import { NextRequest } from 'next/server'
 import { getSavedAuthToken, clearSavedAuthToken } from './api-hooks/auth'
 
-export const baseUrl = 'http://localhost:3000/api'
-//export const baseUrl = 'https://thebigcodetheory.vercel.app/api'
+//export const baseUrl = 'http://localhost:3000/api'
+export const baseUrl = 'https://thebigcodetheory.vercel.app/api'
 
 export const fetchApi = async (input: any, options: any = {}) => {
   const headers = options.headers || {}
